Fix parsing of cosmetic items
Previously, we checked forgedSchematics in the savefile to find out if a
player has unlocked a cosmetic. Past season cosmetics could be obtained
through their seasons performance pass and if you unlock
a cosmetic this way, it wont show up in forgedSchematics (because you
didnt actually 'forge' it).

Now, we check the Vanity section in the savefile to find out if a miner
has access to this vanity item.