Adjust displaying of overall progress and partial progress
We do not want to display a progress of 100% when not all items are
acquried, so we floor the progress. The sum of the progress and partial
progress should also not be at 100% if not all items are at least
unforged. Therefore we floor the progress and ceil the partial progress.